requirements.txt
moviepy>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.28.0
PyYAML>=6.0
schedule>=1.2.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth>=2.23.0
google-cloud-storage>=2.14.0
rembg>=2.0.50
onnxruntime>=1.16.0
scipy>=1.10.0
notion-client>=2.2.0
